Add quit commands and question notifications
With the addition of quit (with checking) Question has been added.
Question stores the command position and the message generated, for
reference by future commands (h and H for example).
The specification does not mention that the ? notifications need to be
recent - it only refers to 'the most recent' even if this is not a
recent command